requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.30
openai>=1.0.0
prompt-toolkit>=3.0.0
rich>=13.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
arxiv>=1.4.0
scholarly>=1.7.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
aiohttp>=3.8.0
aiofiles>=23.0.0
pathlib>=1.0.0
pygments>=2.15.0